cleanup core module
