+++ backend/src/core/__init__.py
@@ -3,9 +3,10 @@
 This file exports the core functions needed to run NSO Bridge.
 """
 
-from .protocols import Memento
+from .memento import Memento
 from .router import api_router, assets, pages_router
-from .schemas import APIResponseClass, CacheKey, ClientSchema, ServerSchema
+from .schemas.api import APIResponseClass
+from .schemas.base import ClientSchema, ServerSchema
 from .service import (
     configure_logging,
     error_handlers,
@@ -13,6 +14,7 @@ from .service import (
     get_server,
     shutdown,
 )
+from .types import CacheKey
 from .utils import (
     endpoint_profiling_middleware,
     get_resource_path,
