+++ backend/src/core/memento.py
