+++ backend/src/core/types.py
@@ -1,0 +1,6 @@
+"""Types."""
+
+from typing import Any, Sequence
+
+type CacheKey = tuple[Any, ...]
+type CacheServerSchema = Sequence[CacheKey]
